requirements.txt
aiofiles==0.8.0
aiosqlite==0.17.0
beautifulsoup4==4.11.1
discord.py==1.7.3
-U git+https://github.com/Rapptz/discord.py
Pillow==9.2.0
requests==2.26.0
selenium==4.3.0
